以下の関数の型を修正 ・dp::newDisplay() ・dp::getWidth( const dp::Display & ) ・dp::getHeight( const dp::Display & ) ・dp::getX() ・dp::setX() ・dp::getY() ・dp::setY() ・dp::getWidth( const dp::DisplayMode & ) ・dp::getHeight( const dp::DisplayMode & )

+++ impl/inc/dpimpl/common/display/display.h
@@ -10,11 +10,11 @@ namespace dp {
     {
         DisplayModeKeyUnique    modeKeyUnique;
 
-        ULong   width;
-        ULong   height;
+        Int width;
+        Int height;
 
-        Long    x;
-        Long    y;
+        Int x;
+        Int y;
 
         DisplayRotate   rotate;
     };
